Move platform::stdio into its own directory.
