+++ src/platform/stdio/windows.rs
