+++ src/platform/stdio/word_wrap.rs
